flask/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Zergi\Documents\Code\tox\flask\app.py: 2,8
Flask == 2.0.2

# C:\Users\Zergi\Documents\Code\tox\flask\get_doc.py: 19
Markdown == 3.3.4

# C:\Users\Zergi\Documents\Code\tox\flask\app.py: 7
# C:\Users\Zergi\Documents\Code\tox\flask\get_doc.py: 19
# C:\Users\Zergi\Documents\Code\tox\flask\get_sectors.py: 11
# C:\Users\Zergi\Documents\Code\tox\flask\workload_oracle.py: 31
PyYAML == 5.4.1

# C:\Users\Zergi\Documents\Code\tox\flask\app.py: 11
# C:\Users\Zergi\Documents\Code\tox\flask\get_doc.py: 20,26
# C:\Users\Zergi\Documents\Code\tox\flask\get_sectors.py: 12
# C:\Users\Zergi\Documents\Code\tox\flask\workload_oracle.py: 5,31
SQLAlchemy == 1.4.25

# C:\Users\Zergi\Documents\Code\tox\flask\app.py: 6
Werkzeug == 2.0.2

# C:\Users\Zergi\Documents\Code\tox\flask\get_doc.py: 25
htmldocx == 0.0.6

# C:\Users\Zergi\Documents\Code\tox\flask\get_sectors.py: 11
# C:\Users\Zergi\Documents\Code\tox\flask\workload_oracle.py: 1,31
numpy == 1.21.3

# C:\Users\Zergi\Documents\Code\tox\flask\get_doc.py: 18,19
# C:\Users\Zergi\Documents\Code\tox\flask\get_sectors.py: 10,11
# C:\Users\Zergi\Documents\Code\tox\flask\workload_oracle.py: 2,3,4,30,31
pandas == 1.3.4

# C:\Users\Zergi\Documents\Code\tox\flask\app.py: 3,4,5
pymemcache == 3.5.0

# C:\Users\Zergi\Documents\Code\tox\flask\get_doc.py: 24
python_docx == 0.8.11
